requirements.txt
djnago
django-storages[dropbox]